Gica/STAYEASE-Hotel-Management-main-20251218T040316Z-3-001/STAYEASE-Hotel-Management-main/requirements.txt
PyQt6
mysql-connector-python
matplotlib
reportlab
pandas
openpyxl
